first working run after refactoring
